Refactor: Move the `transform.rs` module into `transforms` submodule
