+++ projects/api/dxt-lossless-transform-bc1-api/src/transform/builder.rs
@@ -1,8 +1,8 @@
 //! Builder pattern implementation for BC1 estimate options.
 
 use dxt_lossless_transform_api_common::estimate::SizeEstimationOperations;
+use dxt_lossless_transform_bc1::Bc1EstimateOptions;
 use dxt_lossless_transform_bc1::Bc1TransformSettings;
-use dxt_lossless_transform_bc1::transform::Bc1EstimateOptions;
 use dxt_lossless_transform_common::color_565::YCoCgVariant;
 
 /// Builder for BC1 estimate options with convenient configuration methods.
